% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grahespackeges.R
\name{grahespackeges}
\alias{grahespackeges}
\title{Title}
\usage{
grahespackeges(data, x, y, group_by, color_by, plot_type = "line_point")
}
\arguments{
\item{data}{Le jeu de données}

\item{x}{La variable à utiliser sur l'axe des x.}

\item{y}{La variable à utiliser sur l'axe des y .}

\item{group_by}{La variable pour regrouper les données.}

\item{color_by}{La variable pour la couleur des données.}

\item{plot_type}{Le type de graphique à générer.}
}
\value{
Un objet ggplot contenant le graphique.
}
\description{
Description
Générer des graphiques pour différentes variables
}
\details{
Cette fonction crée des graphiques pour différentes variables avec des options de regroupement, de couleur et de type de graphique.
}
\examples{
library(votre_package)
grahespackeges(data, x = "Sepal.Length", y = "Sepal.Width", group_by = "Species", color_by = "Species", plot_type = "line_point")

grahespackeges(data=iris)
}
